+++ builders/codecs/__init__.py
@@ -1,0 +1,1 @@
+"""Codec component builders."""