+++ builders/graphics/__init__.py
@@ -1,0 +1,1 @@
+"""Graphics and filtering component builders."""